Add saturated constructors as values; reorganize
Since saturated constructor calls could also be represented (as before) using
App continuations, some use of smart constructors is now sometimes necessary.
In particular, translation should use them, so now Syntax depends on Ops. But
this is circular. Translation should be in its own module anyway, so I moved it
to the new Translate module. Since this left Ops rather bare, I folded it into
Syntax. But now Syntax and Translate depend on each other, but only in a silly,
hacky way: We do some operations in Syntax (such as computing the type of a
command) by translating to Core and using Core facilities. Since this is a
crutch anyway, I decided to use GHC's bootstrap mechanism (the SOURCE pragma) to
make the modules interdependent.