Reformulate Sequent Core syntax
Let-bound expressions and arguments are now Values, and Commands are embedded as
Values using the Compute constructor. This is simpler in many places; for
instance, it was always a bit hinky to use a type-returning command as a
function argument.

Also, continuations are no longer lists of Frames. The Frame constructors are
now Cont constructors that operate in a chain. This will facilitate, for
instance, using covariables, since a continuation no longer necessarily ends in
an invocation of the One True Covariable.