Use explicit continuation variables everywhere
It arose that the previous system of tracking an implicit continuation variable
wasn't going to work to handle join points the way we wanted. Now we hew closer
to the original sequent calculus, which means explicit bindings for all
continuation variables, namely in the Lam and Compute constructors. Also, as in
the theory, there is no longer an outer continuation "after" a Case; should one
seem to be needed, it can always be simulated by binding a local continuation
variable with the rest of the computation. In fact, this is precisely what the
existing simplifier tries to do anyway.

There are other virtues: A Return is now much more informative than before,
since the variable caches type information; conceivably, an integrated Sequent
Core could hang other data such as whether the continuation represents a
"boring stop" or an "interesting stop," as the GHC simplifier puts it.

That said, the drawbacks are notable: Complexity has increased, both of this
library and of Sequent Core terms, and there are more identifiers running around
taking up memory. Many routines, including the translators and simplifier, now
have to remember the "return id"---essentially, the name of what would have
been the implicit continuation before. It's quite conceivable we'll want to
consider resimplifying, though that would make it all the more necessary to keep
the change to Case so that it does not need to bind the return id.