Split let-bound continuations
Let-bound continuations now get split into an outer continuation that we freely
inline and an inner continuation that we don't. Also, recognize a
case-of-known-constructor pattern when the continuation is let-bound to a case.
Between the two of these, we achieve much of what GHC does with mkDupableCont
and the case-of-case transform, without actually ever performing case-of-case.