Remove the type argument to Case
This argument was only really there out of deference to the original Core type,
but it's pretty meaningless in a sequent core setting: The body of a case
doesn't *have* a type. It was always the type of the linear continuation
variable that lurks in every continuation's context. It's never necessary to
put it here; I guess it could be a check, but then why only annotate Case that
way?

Anyway, getting rid of it made things simpler in most cases and caught at least
one bug (where we should have been using the type of the case binder instead).
Only tricky thing is reconstructing the type when translating back to Core.