Add Lint; fix bugs in handling continuation ids
Given these bug fixes, we can now compile the whole "imaginary" section of the
nofib suite using -O and the Sequent Core simplifier. The only catch is the need
to turn the (new) linter *off* to compile kahan, as it finds a spurious type
error. (It may be either that the linter is erroneous or that the error gets
hidden in translation back to Core.)

The new linter (i.e. type checker for Sequent Core) has already been handy for
pinpointing faults, as it can sometimes detect whether it is the simplifier or
the translation routines that are to blame for an error.