Honor (a bit) unfoldings from Core
This meant adding analogues of DFunUnfolding (the unfolding of a dictionary
value or function) and UnfWhen (guidance that something should usually be
inlined) to the Definition and Guidance types, and thus some more logic in the
inliner.

I'm seeing some errors from the (new) linter that I suspect have to do with the
Float In pass, so dealing with bindings is next. In particular, abusing the tags
of the ids seems not merely ugly but untenable, since after we translate to
Core, something might change the tag using uniqAway.