Mark continuation ids using types, not tags
We now see what ids are for continuations by checking types, rather than looking
at the tag on the unique. This is infinitely nicer and likely more realistic;
if any pass had felt the need to uniqAway a binder, for instance, it would've
replaced the tag. Instead, now we use a type synonym (for (->)) on the Core
side. Also, if a pass makes a continuation no longer syntactically a
continuation, we fall back to making it a regular function.

Also fixed some scope issues (too many things in the InScopeSet).