Fix scope of case binder
Using the wrong environment was letting the scope of a case binder escape the
case statement itself, causing trouble when wildcards become nested. Consider:

case foo wild_00 as wild_00 of { DEFAULT -> wild_00 }

The case binder wild_00 should not capture the argument to foo. It was because
the top-down traversal of the case was passing the environment with the case
binder added to both the alts and the scrutinee.