Make lint check proper scope of continuation variables
Continuation variables should not be used for extraneous control flow
out of a term. Lint now checks that each term contains no reference to
a free continuation variable, which is necessary to ensure that a
Sequent Core term corresponds to a Core term. See the related
documentation about the types of Sequent Core for more details.