Continuation lambdas
A continuation can now be a fixed-arity lambda from some terms to a command.
This is the representation of a join point. This fixes issues with trying to
call join points using unboxed tuples---for one, the simplifier would try to do
case-of-case sorts of things to it, and also unboxed tuples can't have
existential types. (In principle, they could, but they don't currently in GHC.)

Along the way, changed the contification strategy in the translation: Now, we do
a preprocessing step, transforming Core to Core and annotating each let binding
with whether it should translate as a continuation and, if so, what its arity
should be. The decision is made by a bottom-up escape analysis similar to (which
is to say, based on) the one in CoreToStg. The real Core-to-Sequent Core pass
then uses the annotations as it goes; no cleanup pass is needed.