Occurrence analyser
Mostly a copy-and-paste job from GHC's OccurAnal.lhs. At the moment, we use the
new occurrence analyser in the simplifier (thus saving a round-trip
translation each iteration), but we don't exploit its particular advantages.
For instance, we can now see as soon as we traverse a Compute binder whether the
continuation will need to be split, which should let us decide right away
whether we'll ever need to call mkDupableKont.