Split continuations in two
In recently adding polymorphic, polyadic continuations, we went the path of
least resistence and added a "lambda" constructor to continuations and an
"argument tuple" constructor to terms. However, this is an extremely loose
grammar, as we intend that both of these constructs only ever show up at the
top of a command (though perhaps after bindings).

Now, instead we recognize that there are two kinds of continuations: the normal
ones that correspond to strict single-hole evaluation contexts in Core, and
these polyadic ones that represent join points (which we now call parameterized
continuations or PKonts in the code). On the other side of the cut, there are
two kinds of terms: a single expression that gets cut against a regular
continuation, and a tuple that gets cut against a parameterized continuation. So
really, there are two kinds of cuts---term versus continuation and argument
tuple versus parameterized continuation id. (We allow parameterized
continuations to be referred to only by id, thus saving us from needing another
command form or another datatype.)

Changes to the basic datatypes are as follows:

    * The new constructors KLam and KArgs, in Kont and Term respectively, are
      gone.
    * The new PKont type pairs a list of binders with a command. The new Var
      synonym, PKontId, stands for binders for PKonts; they can be distinguished
      from KontIds by having a type like Cont# (ContArgs# _).
    * Command now has three constructors---instead of the original Command
      [Bind] Term Kont, we have Let Bind Command, Eval Term Kont, and Jump
      [Term] PKontId.
    * In the Bind datatype, BindKont is now BindPKont and takes a PKont as its
      second argument. Regular Konts can no longer be let-bound (which actually
      makes the simplifier a bit easier to reason about in places).

Life got a little more complicated for the simplifier. Since there are now two
different sorts of continuations with separate identifier classes, the original
continuation-variable substitution is split into a
parameterized-continuation-variable substitution and a single Maybe for the one
regular continuation that may be bound at once.