Update specification for jumps
In tightening the syntax for Sequent Core, there are now two primary
different commands: evaluating a term in the eye of a continuation,
and jump (with some term arguments) to a let-bound continuation.
Let-bound continuations can now only be of a parameterized sort (a
continuation that accepts zero or more arguments before doing some
work) which can only be referenced by name.  This change makes the
status of join points in the language more clear.