Use mkDupableKont on Compute of Once*
Now that we have our own occurrence analyser, we have at hand the knowledge of
how many places a mu-bound continuation variable is used. This means that when
we simplify a mu-redex (i.e. Eval (Compute _ _) _), we can immediately decide
whether to substitute the continuation whole or whether to avoid duplication by
running it through mkDupableKont first. This makes the whole process very
similar conceptually to the usual inline decision sequence (only we must always
either pre-inline or post-inline), and eliminates a constructor from SimplEnv
that was there to keep track of whether we'd made the continuation dupable yet.