Make argument of coercion "escape"
The escape analysis is a bit of a misnomer: What we're interested in isn't
actually *quite* the same as whether a variable escapes. Case in point:

    f 1 3 `cast` g

The call to f appears to be a tail call, but it isn't because the value it
returns has to be cast by g. However, f still doesn't escape in any meaningful
sense. So as of this commit, the escape analysis lies - it says f escapes, so
that we don't contify it.

The nasty thing is, this is a case where CoreToStg *would* contify (that is,
let-no-escape), because of course it's erasing types, so now f *is* tail-called.
Hopefully this doesn't hurt us much in practice.