Fix certain type parameters when contifying
Suppose we are contifying the polymorphic function:

    k :: forall a b. Bool -> a -> b -> [b]

In contifying k, we give its body a fixed continuation bound in an outer scope;
as b is bound in the body of k, then, the continuation cannot have a type
mentioning b. Hence the contified version of k cannot abstract over b after all.
Now, suppose the outer continuation has type [Int]. Then every invocation of k
must have passed Int as b, since they are all tail calls with that same outer
continuation - otherwise we wouldn't be contifying k. So b has a constant value
to begin with, and we can substitute. The contification process then results in
a parameterized continuation

    k :: Cont# (exists a. (Bool, a, Int))

(type simplified for clarity). Note that since a doesn't appear in the original
function's return type, it is free to vary, and we construct the existential as
usual. This is important for case analyses on existential types, which produce
polymorphic join points.