Remove (unparameterized) continuation ids
Since there was only ever one KontId in scope at any point, it was always
redundant to put one as an argument to Compute or Return, besides to have a
convenient place to put type and occurrence information. We now put the type in
the Compute form (as we did some eons ago) instead of a KontId; the occurrence
information was not crucial, and we once again wait to make the continuation
dupable until we hit a multi-way case statement. (We anticipate generalizing the
idea and uses of the current continuation rather dramatically, so that the
occurrence info was probably not going to remain useful anyway.)