Refactor main simplifier sequence; implement termIsConApp_maybe
Rather than disappearing into an inner loop inside simplKont, execution through
the main simplifier now remains in top-level functions and consistently proceeds
by tail recursion. This should help when we start messing with the control flow
with things like metacontinuations.

Also, we now delay finding a known branch until after we've simplified the whole
scrutinee (i.e. the term and all frames), much as GHC does, and we have a
version of exprIsConApp_maybe to search for a known branch.