Contification: Deal with recursion+polymorphism
Previously it was believed that, if we need to fix a type argument of a
continuation, we could just look at any of the invocations of the original
function. This only works for non-recursive functions; for recursive functions,
some of the calls are (of course) recursive, and these will simply pass along
the type variable. So now we keep track of which calls are recursive and look
at an "outside" call when there is one; when there isn't (perhaps the function
is mutually recursive and *never* invoked from the outside), we give up on
contification. It may be possible to support this case as well, but it would
require significant effort - either unification or abstract interpretation
AFAICT.