Metacontinuations and strict arguments
Since we don't have a continuation for applying a result as a strict argument,
we need to generalize the idea of the current continuation stored in the
environment in order to simplify strict arguments in a call-by-value manner.
Enter metacontinuations: Now the environment can hold a context that's either
a syntactic continuation *or* a continuation for the simplifier itself. We
defunctionalize the continuation, storing the arguments to pass to simplKont, so
that mkDupableKont can inspect and modify it.