Implement rewrite rules
The tricky part here isn't the rewrite itself---we just have to translate to
Core before calling lookupRule and then translate the result back. The bigger
problem is what to do with leftover frames; we already simplified them, so we
don't just want to tack them back onto the end of the InKont we've been working
on. The solution is a new kind of metacontinuation that simply appends
already-simplified frames without further processing.