Fix handling of coercions
We now distinguish between when we need to *cast* coercions for type safety and
when we actually *optimize* them, which we still put off as long as we can.
Important because we would like to run a SynKont metacontinuation early in order
to pull in extra arguments, but that means the ArgInfo's coercion must already
be an OutCoercion as we're about to lose its environment.

Also, simplified the chaining of metacontinuations by adding an InEnd field to
ApplyOutFrames, eliminating the need to create a separate SynKont for just the
end.