Strict arguments to jumps
Jump arguments were always being evaluated as if lazy. Fixing this without
duplicating tons of code meant hijacking simplKont so that it's not necessarily
actually simplifying a continuation, it may be simplifying a list of application
frames that's going to be the argument list of a Jump or bound in the body of a
PKont. We could make this less ugly by changing the representation of jumps
(again!) to make Jump an End.