Translate: Share cont using join point, not Compute
The scope rules require that the body of a Compute be continuation-closed, so we
can't always use it to bind a continuation for sharing purposes. Instead we
create a simple join point. We have to use a case as a shim, but since
continuations are strict contexts, this should (!) not cause extra strictness.