Excise Floats from SimplEnv
Floats have impact on the environment because the in-scope set needs to be
augmented with the binders being floated, but otherwise keeping them together is
conceptually muddy. Now that floats are returned separately from environments,
there is a type distinction between functions that float things *and add
bindings to the current environment* and those that merely generate floats:

  simplTerm :: SimplTermEnv -> CallCtxt -> InTerm -> SimplM (Floats, OutTerm)
    -- Returns a term; may also generate floats
  simplTermNoFloats :: SimplTermEnv -> CallCtxt -> InTerm -> SimplM OutTerm
    -- Returns a term; never generates floats
  simplNonRecOut :: SimplEnv -> InId -> OutTerm -> SimplM (Floats, SimplEnv)
    -- Returns an augmented environment; may also generate floats

It is an invariant that functions that return both Floats *and* a new
environment must have already updated the new environment with the floated
bindings.