Got rid of Rhs type
This was largely only providing a false sense of security in making it look like
terms and joins are often dealt with the same way. They are sometimes (the
pre-inlining decision looks pretty much the same, for instance), but more often
what was happening was that I would unthinkingly create rhsIsCheap and
joinIsCheap to go with termIsCheap and not realize that the cheap predicate is
always true of join points (as they are effectively always lambdas). Type safety
is also somewhat improved.

If we want to be able to project the RHS out of BindPair, we could always make
BindPair a GADT ...