Normalize coercions ahead of time
The logic for dealing with coercions is rather complex and diffused
throughout the simplifier - all the functions in the main loop take a
Maybe SubstedCoercion (either separately or as part of an ArgInfo) so that they
can, for instance, implement the Push rule to push a cast into and past an
argument application. At Simon's suggestion, we can just do what the original
simplifier does and normalize the continuation at an early stage, then not worry
about clever coercion tricks from then on. This simplifies a lot of logic by
coalescing all the coercion stuff into the new normalizeKont function.