Split SimplEnv
Gone are the days of keeping fragments of the environment around and
splicing them in as needed; now we just have separate datatypes for those
different fragments. The first step of this was really when we separated Floats
from the environment. Now all the substitutions are also now gone, as is the
StaticEnv newtype that served as a mask for those fields that constituted
lexically-scoped information. Now those fields all live in either DataScope or
ControlScope; these are kept separate because many objects, including terms and
frames, never "see" bound continuations (neither "ret" nor join points). Types
now make it much clearer what's going on; simplTerm, for instance, simply does
not take a ControlScope for its argument. We also don't need silliness like
setStaticPart to combine a lexical environment with an in-scope set; we simply
pass the old in-scope set and the new lexical environment to whatever function
we were calling. It's unclear how much this saves us in runtime (maybe W/W was
undoing the copying overhead?), but it's definitely conceptually cleaner.

(Incidentally, "data scope" and "control scope" are just the terms I came up
with now; in particular we might use "control variable" to refer collectively to
join points and the continuation binding. But since they're new to our
vocabulary we can change them easily.)